scripts/requirements.txt
# OCR pipeline dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.24.0
pyodbc>=5.1.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
